+++ src/lib/lessWatchCompilerUtils.cjs.js
@@ -89,72 +89,6 @@ define(function (require) {
         },
         initCallback);
     },
-    // Here's where we run the less compiler
-    getLessArgs: function(args) {
-      var arr = args.split(',');
-      return " --" + arr.join(' --');
-    },
-    compileCSS: function (file, test) {
-
-      var outputFilePath = this.resolveOutputPath(file);
-
-      // As a rule, we don't compile hidden files for now. If we encounter one,
-      // just return.
-      if (fileSearch.isHiddenFile(outputFilePath)) return
-
-      var enableJsFlag = lessWatchCompilerUtilsModule.config.enableJs ? ' --js' : '';
-      var minifiedFlag = lessWatchCompilerUtilsModule.config.minified ? ' -x' : '';
-      var sourceMap = (lessWatchCompilerUtilsModule.config.sourceMap) ? ' --source-map' : '';
-      var lessArgs = (lessWatchCompilerUtilsModule.config.lessArgs)? this.getLessArgs(lessWatchCompilerUtilsModule.config.lessArgs): '';
-      var plugins = (lessWatchCompilerUtilsModule.config.plugins) ? ' --' + lessWatchCompilerUtilsModule.config.plugins.split(',').join(' --') : '';
-      var command = 'lessc' + lessArgs + sourceMap + enableJsFlag + minifiedFlag + plugins + ' ' + JSON.stringify(file) + ' ' + outputFilePath;
-      // Run the command
-      //  console.log(command)
-      if (!test)
-        exec(command, function (error, stdout, stderr) {
-          if (error !== null) {
-            console.log(error);
-            if (lessWatchCompilerUtilsModule.config.runOnce)
-              process.exit(1);
-          }
-          if (stdout)
-            console.error(stdout);
-        });
-      return {
-        "command": command,
-        "outputFilePath": outputFilePath
-      };
-
-    },
-    resolveOutputPath: function (filePath) {
-      var fullPath = path.resolve(filePath);
-      var parsedPath = path.parse(fullPath);
-
-      // Only empty when unit testing it seems
-      var relativePath = null;
-      var dirname = null;
-      if (lessWatchCompilerUtilsModule.config.watchFolder) {
-        relativePath = path.relative(lessWatchCompilerUtilsModule.config.watchFolder, fullPath);
-        dirname = path.dirname(relativePath);
-      } else {
-        dirname = path.dirname(filePath);
-      }
-      var filename = parsedPath.name;
-
-      var formatted = path.format({
-        dir: dirname,
-        name: filename,
-        ext: (lessWatchCompilerUtilsModule.config.minified ? '.min' : '') + '.css',
-      });
-
-      // No matter the path of the main file, the output must always land in the output folder
-      formatted = formatted.replace(/^(\.\.[\/\\])+/, '');
-
-      var finalFullPath = path.resolve(lessWatchCompilerUtilsModule.config.outputFolder, formatted);
-      var shortPath = path.relative(cwd, finalFullPath);
-
-      return JSON.stringify(shortPath);
-    },
     // We build the function to filter the files to watch.
     // Returning true marks a file to be ignored.
     filterFiles: function (f) {
